Support for current wxpython started

+++ tests/big_import.py
@@ -1,5 +1,6 @@
 """Module which will require a lot of time to import"""
+from six.moves import range
 
 for i in range( 2500 ):
-	long(i) ** i
+	int(i) ** i
